Create ActiveSupport type
