+++ lib/awesome_print/types.rb
@@ -1,5 +1,7 @@
 module AwesomePrint
   module Types
+    require 'awesome_print/types/base'
     require 'awesome_print/types/active_record' if defined?(ActiveRecord) || AwesomePrint.rails_console?
+    require 'awesome_print/types/active_support' if defined?(ActiveSupport) || AwesomePrint.rails_console?
   end
 end
