+++ lib/awesome_print/types/base.rb
@@ -1,0 +1,14 @@
+module AwesomePrint
+  module Types
+    class Base
+
+      def initialize(object)
+        @object = object
+      end
+
+      private
+
+        attr_reader :object
+    end
+  end
+end
